--- main.css
@@ -1,86 +1,0 @@
-.o-modal-body-show {
-  overflow: hidden;
-}
-
-.is-mobile .o-modal-body-show {
-  overflow: visible;
-}
-
-.o-modal {
-  top: 0;
-  left: 0;
-  width: 100%;
-  text-align: center;
-  display: none;
-}
-
-.o-modal.o-modal-show {
-  display: block;
-}
-
-.is-not-mobile .o-modal {
-  position: fixed;
-  height: 100%;
-  white-space: nowrap;
-}
-
-.is-not-mobile .o-modal::before {
-  content: '';
-  display: inline-block;
-  height: 100%;
-  vertical-align: middle;
-  margin-right: -0.25em;
-}
-
-.is-not-mobile .o-modal:before {
-  content: '';
-  display: inline-block;
-  height: 100%;
-  vertical-align: middle;
-  margin-right: -0.25em;
-}
-
-.is-mobile .o-modal {
-  position: absolute;
-}
-
-.o-modal-backdrop {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 100%;
-  filter: alpha(opacity=50);
-  background-color: black;
-  opacity: 0.5;
-}
-
-.o-modal-positioner {
-  display: inline-block;
-  position: relative;
-  text-align: left;
-  white-space: normal;
-}
-
-.is-not-mobile .o-modal-positioner {
-  vertical-align: middle;
-}
-
-.o-modal-close {
-  -webkit-appearance: none;
-  padding: 0;
-  cursor: pointer;
-  background: none;
-  border: 0;
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  width: 21px;
-  font-size: 21px;
-  line-height: 1;
-  font-weight: 700;
-  color: black;
-  text-shadow: 0 1px 0 white;
-  filter: alpha(opacity=20);
-  opacity: 0.2;
-}
