+++ package.json
@@ -14,11 +14,19 @@
     "modal",
     "mobile"
   ],
-  "author": "",
+  "author": "Daiwei Lu <daiweilu123@gmail.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/spanishdict/supermodal/issues"
   },
   "homepage": "https://github.com/spanishdict/supermodal",
-  "devDependencies": {}
+  "devDependencies": {
+    "browser-sync": "^2.0.0-rc1",
+    "browserify": "^8.1.1",
+    "gulp": "^3.8.10",
+    "gulp-sass": "^1.3.2",
+    "gulp-watch": "^4.1.0",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.0.0"
+  }
 }
